jsonpath/src/main/java/io/github/xmljim/json/jsonpath/function/node/NodeTypeFunction.java
package io.github.xmljim.json.jsonpath.function.node;

import io.github.xmljim.json.jsonpath.context.Context;
import io.github.xmljim.json.jsonpath.function.AbstractJsonPathFunction;
import io.github.xmljim.json.jsonpath.function.info.FunctionDefinition;
import io.github.xmljim.json.jsonpath.util.BuiltIns;
import io.github.xmljim.json.jsonpath.util.Global;

import java.util.Collections;
import java.util.Locale;
import java.util.stream.Stream;

@FunctionDefinition(builtIn = BuiltIns.TYPE)
public class NodeTypeFunction extends AbstractJsonPathFunction {
    public NodeTypeFunction(Global global) {
        super(BuiltIns.TYPE.functionName(), Collections.emptyList(), global);
    }

    @Override
    public Stream<Context> apply(Stream<Context> contextStream) {
        return contextStream.map(context -> Context.createSimpleContext(context.type().name().toLowerCase(Locale.ROOT)));
    }
}
